backend/requirements/dev.txt
-r prod.txt
django-cors-headers>=4.4
pytest>=8.2
pytest-django>=4.9
factory_boy>=3.3